depression/EEG/requirements_windows_lsl.txt
bleak==2.1.1
pyserial==3.5
pylsl==1.17.6
python-dotenv==1.1.1
